split build tests
